requirements.txt
pandas~=2.2.3
numpy~=2.2.5
statsmodels~=0.14.4
torch~=2.7.0
scikit-learn~=1.6.1
PyYAML~=6.0.2
matplotlib~=3.10.3
seaborn~=0.13.2
shap~=0.40.0
faiss-cpu~=1.7.0